pyproject.toml
[tool.poetry]
name = "python-selenium"
version = "0.1.0"
description = "Demo of a Python Selenium project"
authors = ["Joe"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
pillow = "^9.5.0"
pytesseract = "^0.3.10"
easyocr = "^1.7.0"

[tool.poetry.dev-dependencies]
selenium = "^4.10.0"
pytest = "^7.4.0"
webdriver-manager = "^3.9.1"
ddt = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest-html = "^3.2.0"
pytest-xdist = "^3.1.0"
openpyxl = "^3.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
